understanding the code
