+++ react-native-shopping-app/src/constant/data.ts
@@ -1,0 +1,91 @@
+
+const MESONARY_LIST_DATA = [
+  {
+    imageUrl:
+      "https://images.unsplash.com/photo-1521577352947-9bb58764b69a?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=986&q=80",
+    title: "PUMA Everyday Hussle",
+    price: 160,
+  },
+  {
+    imageUrl:
+      "https://images.unsplash.com/photo-1586790170083-2f9ceadc732d?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=987&q=80",
+    title: "PUMA Everyday Hussle",
+    price: 180,
+  },
+  {
+    imageUrl:
+      "https://images.unsplash.com/photo-1556217477-d325251ece38?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1020&q=80",
+    title: "PUMA Everyday Hussle",
+    price: 200,
+  },
+  {
+    imageUrl:
+      "https://images.unsplash.com/photo-1554568218-0f1715e72254?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=987&q=80",
+    title: "PUMA Everyday Hussle",
+    price: 180,
+  },
+  {
+    imageUrl:
+      "https://images.unsplash.com/photo-1627225924765-552d49cf47ad?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=987&q=80",
+    title: "PUMA Everyday Hussle",
+    price: 120,
+  },
+  {
+    imageUrl:
+      "https://images.unsplash.com/photo-1586790170083-2f9ceadc732d?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=987&q=80",
+    title: "PUMA Everyday Hussle",
+    price: 180,
+  },
+  {
+    imageUrl:
+      "https://images.unsplash.com/photo-1556217477-d325251ece38?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1020&q=80",
+    title: "PUMA Everyday Hussle",
+    price: 200,
+  },
+  {
+    imageUrl:
+      "https://images.unsplash.com/photo-1554568218-0f1715e72254?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=987&q=80",
+    title: "PUMA Everyday Hussle",
+    price: 180,
+  },
+  {
+    imageUrl:
+      "https://images.unsplash.com/photo-1521577352947-9bb58764b69a?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=986&q=80",
+    title: "PUMA Everyday Hussle",
+    price: 160,
+  },
+  {
+    imageUrl:
+      "https://images.unsplash.com/photo-1586790170083-2f9ceadc732d?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=987&q=80",
+    title: "PUMA Everyday Hussle",
+    price: 180,
+  },
+  {
+    imageUrl:
+      "https://images.unsplash.com/photo-1556217477-d325251ece38?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1020&q=80",
+    title: "PUMA Everyday Hussle",
+    price: 200,
+  },
+  {
+    imageUrl:
+      "https://images.unsplash.com/photo-1554568218-0f1715e72254?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=987&q=80",
+    title: "PUMA Everyday Hussle",
+    price: 180,
+  },
+];
+
+const CATEGORIES = [
+  "Clothing",
+  "Shoes",
+  "Accessories",
+  "Accessories 2",
+  "Accessories 3",
+  "Accessories 4",
+];
+
+const data = {
+  MESONARY_LIST_DATA,
+  CATEGORIES
+}
+
+export default data
